Fish_Weight_Prediction
With a dataset of fish species, with some of it characteristic like it vertical, diagonal, length, height, and width. I will try to predict the weight of the fish based on their characteristic. I will use Linear Regression Method to see whether the weight of the fish related to their characteristic.